+++ src/main/java/com/sistema/examenes/services/MetaPDOT_ServiceImpl.java
@@ -42,7 +42,7 @@ public class MetaPDOT_ServiceImpl extends GenericServiceImpl<MetaPDOT, Long> imp
 
         for (Object[] resultado : resultados) {
             MetaPdot_DTO metaDTO = new MetaPdot_DTO(
-                    (Long) resultado[0],
+                    ((BigInteger) resultado[0]).longValue(),
                     (String) resultado[1],
                     (String) resultado[2],
                     (double) resultado[3]
